Destructured further into discreet submodules
