+++ CancerGov/_src/Scripts/NCI/Modules/liveHelpPopup/settings/index.js
@@ -1,0 +1,15 @@
+import ctsEnglishSettings from './ctsEnglishSettings';
+import ctsSpanishSettings from './ctsSpanishSettings';
+
+const basePaths = [
+    {
+        path: /^\/about-cancer\/treatment\/clinical-trials/,
+        settings: ctsEnglishSettings
+    },
+    {
+        path: /^\/espanol\/cancer\/tratamiento/,
+        settings: ctsSpanishSettings
+    }
+];
+
+export default basePaths;
